+++ trixnity-crypto-driver/trixnity-crypto-driver-libolm/src/commonMain/kotlin/net/folivo/trixnity/crypto/driver/libolm/keys/LibOlmCurve25519PublicKey.kt
@@ -1,0 +1,18 @@
+package net.folivo.trixnity.crypto.driver.libolm.keys
+
+import net.folivo.trixnity.crypto.driver.keys.Curve25519PublicKey
+import net.folivo.trixnity.utils.decodeUnpaddedBase64Bytes
+import kotlin.jvm.JvmInline
+
+@JvmInline
+value class LibOlmCurve25519PublicKey(internal val inner: String) : Curve25519PublicKey {
+
+    override val base64: String
+        get() = inner
+
+    override val bytes: ByteArray
+        get() = base64.decodeUnpaddedBase64Bytes()
+
+    override fun close() {}
+}
+