+++ trixnity-crypto-driver/trixnity-crypto-driver-libolm/src/commonMain/kotlin/net/folivo/trixnity/crypto/driver/libolm/olm/LibOlmMessageFactory.kt
@@ -1,0 +1,8 @@
+package net.folivo.trixnity.crypto.driver.libolm.olm
+
+import net.folivo.trixnity.crypto.driver.olm.MessageFactory
+
+object LibOlmMessageFactory : MessageFactory {
+    override val normal: LibOlmNormalMessageFactory = LibOlmNormalMessageFactory
+    override val preKey: LibOlmPreKeyMessageFactory = LibOlmPreKeyMessageFactory
+}